Stereographic azimuthal projection bug fix
When the projection latitude is either +/-90 degrees, the PROJ4J stereographic azimuthal projection value diverges
from the PROJ4 command line value. The root cause of this divergence is that the true scale latitude was not being
properly updated. To match the command line output, the true scale latitude must be set to the projection latitude
unless a specific true scale latitude is specified via the lat_ts parameter in the PROJ4 string.